Update methods added Exceptions pasted

+++ src/main/java/com/restaurantsProject/project/services/foodService/FoodServiceImpl.java
@@ -2,7 +2,11 @@ package com.restaurantsProject.project.services.foodService;
 
 import com.restaurantsProject.project.entities.Food;
 import com.restaurantsProject.project.entities.Restaurant;
+import com.restaurantsProject.project.exceptions.CouldNotDeleteDataException;
+import com.restaurantsProject.project.exceptions.CouldNotUpdateDataException;
+import com.restaurantsProject.project.exceptions.DataNotFoundException;
 import com.restaurantsProject.project.repositories.FoodRepository;
+import com.restaurantsProject.project.repositories.RestaurantRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,6 +17,9 @@ public class FoodServiceImpl implements FoodService {
 
     @Autowired
     FoodRepository foodRepository;
+
+    @Autowired
+    RestaurantRepository restaurantRepository;
 
 
     @Override
@@ -25,6 +32,9 @@ public class FoodServiceImpl implements FoodService {
 
     @Override
     public List<Food> getFoodsByRestaurant(Restaurant restaurant) {
+        Restaurant gettingRestaurant = restaurantRepository.findRestaurantById(restaurant.getId());
+        if (gettingRestaurant == null)
+            throw new DataNotFoundException();
         List<Food> foodsByRestaurant = foodRepository.findAllByRestaurant(restaurant);
 
         return foodsByRestaurant;
@@ -33,20 +43,34 @@ public class FoodServiceImpl implements FoodService {
     @Override
     public Food updateFood(long id, Food food) {
         Food updatingFood = foodRepository.findFoodById(id);
-        return null;
+        if (updatingFood == null)
+            throw new CouldNotUpdateDataException();
+
+        updatingFood.setIngredients(food.getIngredients());
+        updatingFood.setName(food.getName());
+        updatingFood.setPrice(food.getPrice());
+        updatingFood.setType(food.getType());
+        Food returningFood = foodRepository.save(updatingFood);
+
+        return returningFood;
     }
 
     @Override
     public void deleteFood(long id) {
+        Food deletingFood = foodRepository.findFoodById(id);
+        if (deletingFood == null)
+            throw new CouldNotDeleteDataException();
         foodRepository.deleteFoodById(id);
     }
 
     @Override
     public Food getFood(long id) {
+
         Food food = foodRepository.findFoodById(id);
+        if (food == null)
+            throw new DataNotFoundException();
         return food;
     }
-
 
 
     @Override
@@ -65,7 +89,7 @@ public class FoodServiceImpl implements FoodService {
 
     @Override
     public List<Food> getFoodsByPrice(double lower, double higher) {
-        List<Food> allFoodsByPrice = foodRepository.findAllByPriceBetween(lower,higher);
+        List<Food> allFoodsByPrice = foodRepository.findAllByPriceBetween(lower, higher);
 
         return allFoodsByPrice;
     }
